unify search code on backend (categories + functors)
